+++ bankfeeds/src/codatbankfeeds/account_mapping.py
@@ -53,8 +53,8 @@ class AccountMapping:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.AccountMappingResult])
-                res.account_mapping_result = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.BankFeedAccountMappingResponse])
+                res.bank_feed_account_mapping_response = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [400, 401, 404, 429]: