+++ bankfeeds/src/codatbankfeeds/models/operations/delete_connection.py
@@ -4,29 +4,25 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
-from ..shared import pushoperation as shared_pushoperation
 from typing import Optional
 
 
 
 @dataclasses.dataclass
-class GetPushOperationRequest:
+class DeleteConnectionRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
-    push_operation_key: str = dataclasses.field(metadata={'path_param': { 'field_name': 'pushOperationKey', 'style': 'simple', 'explode': False }})
-    r"""Push operation key."""
+    connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
     
 
 
 
 
 @dataclasses.dataclass
-class GetPushOperationResponse:
+class DeleteConnectionResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
-    push_operation: Optional[shared_pushoperation.PushOperation] = dataclasses.field(default=None)
-    r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)