+++ bankfeeds/RELEASES.md
@@ -759,3 +759,13 @@ Based on:
 - [python v0.36.0] bankfeeds
 ### Releases
 - [PyPI v0.36.0] https://pypi.org/project/codat-bankfeeds/0.36.0 - bankfeeds
+
+## 2023-09-01 00:17:24
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.77.0 (2.91.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.37.0] bankfeeds
+### Releases
+- [PyPI v0.37.0] https://pypi.org/project/codat-bankfeeds/0.37.0 - bankfeeds